Move function is_snapshot_on_nfs from utils to common
