+++ test/integration/testpaths/testpath_snapshot_hadrning.py
@@ -19,31 +19,35 @@
 
 from nose.plugins.attrib import attr
 from marvin.cloudstackTestCase import cloudstackTestCase, unittest
-from marvin.lib.utils import (cleanup_resources,
-                              is_snapshot_on_nfs,
-                              validateList)
-from marvin.lib.base import (Account,
-                             Cluster,
-                             StoragePool,
-                             DiskOffering,
-                             ServiceOffering,
-                             Host,
-                             Configurations,
-                             Template,
-                             VirtualMachine,
-                             Snapshot,
-                             SnapshotPolicy,
-                             Volume
-                             )
-from marvin.lib.common import (get_domain,
-                               get_zone,
-                               get_template,
-                               list_volumes,
-                               list_snapshots,
-                               list_virtual_machines,
-                               createChecksum,
-                               compareChecksum
-                               )
+from marvin.lib.utils import (
+    cleanup_resources,
+    validateList
+)
+from marvin.lib.base import (
+    Account,
+    Cluster,
+    StoragePool,
+    DiskOffering,
+    ServiceOffering,
+    Host,
+    Configurations,
+    Template,
+    VirtualMachine,
+    Snapshot,
+    SnapshotPolicy,
+    Volume
+)
+from marvin.lib.common import (
+    get_domain,
+    get_zone,
+    get_template,
+    list_volumes,
+    list_snapshots,
+    list_virtual_machines,
+    createChecksum,
+    compareChecksum,
+    is_snapshot_on_nfs
+)
 from marvin.sshClient import SshClient
 import time
