+++ test/integration/testpaths/testpath_volume_snapshot.py
@@ -18,26 +18,30 @@
 """
 from nose.plugins.attrib import attr
 from marvin.cloudstackTestCase import cloudstackTestCase, unittest
-from marvin.lib.utils import (cleanup_resources,
-                              is_snapshot_on_nfs,
-                              validateList)
-from marvin.lib.base import (Account,
-                             ServiceOffering,
-                             DiskOffering,
-                             Template,
-                             VirtualMachine,
-                             Snapshot,
-                             Volume
-                             )
-from marvin.lib.common import (get_domain,
-                               get_zone,
-                               get_template,
-                               list_volumes,
-                               list_snapshots,
-                               list_events,
-                               createChecksum,
-                               compareChecksum
-                               )
+from marvin.lib.utils import (
+    cleanup_resources,
+    validateList
+)
+from marvin.lib.base import (
+    Account,
+    ServiceOffering,
+    DiskOffering,
+    Template,
+    VirtualMachine,
+    Snapshot,
+    Volume
+)
+from marvin.lib.common import (
+    get_domain,
+    get_zone,
+    get_template,
+    list_volumes,
+    list_snapshots,
+    list_events,
+    createChecksum,
+    compareChecksum,
+    is_snapshot_on_nfs
+)
 
 from marvin.codes import PASS
 from threading import Thread
